Prepare a page
You can now prepare a page when you're still pregnant.
A button is added on the frontpage for this.
The page will only be visible to you until published.
An register e-mail is send when preparing which differs from the publish e-mail.
The first strep of creating a page is slightly different when preparing.